# util/jsonutil: new package