This significantly cuts allocations for some alloc-hot corp code. That benchmark:

```
name      old time/op    new time/op    delta
TimeOf-8    2.12µs ± 1%    1.96µs ± 0%   -7.36%  (p=0.000 n=9+8)

name      old alloc/op   new alloc/op   delta
TimeOf-8      296B ± 0%       48B ± 0%  -83.78%  (p=0.000 n=10+10)

name      old allocs/op  new allocs/op  delta
TimeOf-8      6.00 ± 0%      1.00 ± 0%  -83.33%  (p=0.000 n=10+10)
```