@josharian 2020-11-19
Just realized I should probably use bytes.Reader and Reset in here, not bytes.Buffer. Will fix tomorrow.